requirements.txt
dj-database-url==2.0.0
django-environ==0.10.0
Django==4.2.2
djangorestframework==3.14.0
gunicorn==20.1.0
joblib==1.2.0
keras==2.12.0
numpy==1.23.5
pandas==2.0.2
psycopg2==2.9.6
requests==2.31.0
scikit-learn==1.2.2
tensorflow==2.12.0
whitenoise==6.5.0
